testExperimentCheckpointWithDatasets
